refactor(mcp-server): drop misleading "global-" prefix on memory files
The memory collection's store and node-routes files were named
global-memory-*, where "global" (machine-wide, not per-project) collides
confusingly with "global context". Rename to memory-store.ts /
memory-node-routes.ts and registerMemoryNodeRoutes(), and move the tests
to tests/memory/. The collection itself is correctly named "memory";
only the misleading prefix is dropped.
