+++ src/mcp-server/mcp-server.ts
@@ -9,14 +9,14 @@ import { VERSION } from "#src/shared/config.ts";
 import { checkForUpdate } from "#src/shared/version-check.ts";
 import { createExpressApp } from "./create-express-app.ts";
 import { registerGlobalContextNodeRoutes } from "./helpers/global-context/global-context-node-routes.ts";
-import { registerGlobalMemoryNodeRoutes } from "./helpers/memory/global-memory-node-routes.ts";
+import { registerMemoryNodeRoutes } from "./helpers/memory/memory-node-routes.ts";
 import { registerCustomSkillsNodeRoutes } from "./helpers/skills-custom/custom-skills-node-routes.ts";
 import { registerLibraryRoutes } from "./live-library/library-routes.ts";
 import * as console from "./node-for-max-logger.ts";
 
 registerLibraryRoutes();
 registerGlobalContextNodeRoutes();
-registerGlobalMemoryNodeRoutes();
+registerMemoryNodeRoutes();
 registerCustomSkillsNodeRoutes();
 
 interface ServerError extends Error {
